chore: code review completo, BigDecimal, Swagger, testes, Postman e docs
Backend:
- Refactor @Data -> @Getter/@Setter nas 14 entidades (evita LazyInitializationException)
- Double -> BigDecimal (precision=10, scale=2) nos campos de dinheiro
- Validacao dataCheckout > dataCheckin em ReservaService (HTTP 400)
- GlobalExceptionHandler agora trata IllegalArgumentException (400) e DataIntegrityViolationException (409)
- Config CORS aberto em WebConfig
- Swagger/OpenAPI via springdoc 2.7.0 + OpenApiConfig (/swagger-ui.html)
- Spring Boot 4 -> 3.4.1 LTS / Java 21 -> 17 (estabilidade)
- application.properties apontando default para PG16 (porta 5433)
- Fix typo no ReservaServicoRequestDTO e ;; duplo no ReservaServicoService

Testes:
- 58 testes JUnit + MockMvc cobrindo os 14 controllers
- H2 in-memory via profile 'test', sem dependencia de PostgreSQL local
- TestDataFactory para criacao de entidades pai em testes com FK

Frontend:
- Escapes XSS consistentes (htmlspecialchars + json_encode em strings JS)

Docs:
- README.md raiz completo (setup, env, troubleshooting, decisoes arquiteturais)
- README-YURI.md atualizado com instrucoes de teste e roteiro do video
- docs/postman/HotelApi.postman_collection.json com 76 requests organizadas
- .gitignore ignorando pasta .claude/

+++ backend/src/main/java/br/com/fatec/hotel/dtos/CargoRequestDTO.java
@@ -1,5 +1,7 @@
 package br.com.fatec.hotel.dtos;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
@@ -19,6 +21,6 @@ public class CargoRequestDTO {
 
     @NotNull(message = "O salário base é obrigatório")
     @Positive(message = "O salário base deve ser positivo!")
-    private Double salarioBase;
+    private BigDecimal salarioBase;
 
 }
