+++ backend/src/main/java/br/com/fatec/hotel/dtos/TipoQuartoResponseDTO.java
@@ -1,5 +1,7 @@
 package br.com.fatec.hotel.dtos;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -10,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class TipoQuartoResponseDTO {
     private Long codTipo;
     private String nome;
-    private Double precoBase;
+    private BigDecimal precoBase;
 }
